fix(auto-reply): split agent event bridge module and fix plan type boundaries
