+++ src/auto-reply/reply/agent-runner-cli-dispatch.ts
@@ -22,91 +22,16 @@ import { buildPlanUpdateStepFields } from "../../channels/streaming.js";
 import type { SessionEntry } from "../../config/sessions.js";
 import { updateSessionEntry } from "../../config/sessions/session-accessor.js";
 import type { AgentEventPayload } from "../../infra/agent-events.js";
-import {
-  emitAgentEvent,
-  onAgentEvent,
-  withAgentRunLifecycleGeneration,
-} from "../../infra/agent-events.js";
+import { emitAgentEvent, withAgentRunLifecycleGeneration } from "../../infra/agent-events.js";
 import { FAST_MODE_AUTO_PROGRESS_KIND, type ReplyPayload } from "../reply-payload.js";
 import { formatToolAggregate } from "../tool-meta.js";
 import type { GetReplyOptions } from "../types.js";
+import {
+  type AgentEventDeliveryStartOrder,
+  createAgentEventBridge,
+  createAgentEventDeliveryStartOrder,
+} from "./agent-event-bridge.js";
 import { resolveAgentLifecycleTerminalMetadata } from "./agent-lifecycle-terminal.js";
-
-type AgentEventDeliveryStartOrder = {
-  schedule: (deliver: () => Promise<void>) => Promise<void>;
-};
-
-function createAgentEventDeliveryStartOrder(): AgentEventDeliveryStartOrder {
-  let startTail = Promise.resolve();
-  return {
-    schedule: async (deliver) => {
-      // Reserve at raw event receipt, then release at callback invocation. CLI streams drain
-      // independently, so waiting for callback completion here would reorder later streams.
-      const previousStart = startTail;
-      let releaseStart: (() => void) | undefined;
-      startTail = new Promise<void>((resolve) => {
-        releaseStart = resolve;
-      });
-      await previousStart;
-      let delivery: Promise<void>;
-      try {
-        delivery = deliver();
-      } finally {
-        releaseStart?.();
-      }
-      await delivery;
-    },
-  };
-}
-
-function createAgentEventBridge<T>(params: {
-  runId: string;
-  suppressed?: boolean;
-  read: (evt: AgentEventPayload) => T | undefined;
-  deliver?: (payload: T) => Promise<void>;
-  startOrder?: AgentEventDeliveryStartOrder;
-}) {
-  const deliver = params.deliver;
-  if (!deliver) {
-    return {
-      unsubscribe: () => undefined,
-      drain: async (): Promise<void> => undefined,
-    };
-  }
-  let unsubscribed = false;
-  let delivery = Promise.resolve();
-  const rawUnsubscribe = onAgentEvent((evt) => {
-    if (evt.runId !== params.runId) {
-      return;
-    }
-    if (params.suppressed) {
-      return;
-    }
-    const payload = params.read(evt);
-    if (payload === undefined) {
-      return;
-    }
-    if (!params.startOrder) {
-      delivery = delivery.then(() => deliver(payload)).catch(() => undefined);
-      return;
-    }
-    const scheduled = params.startOrder.schedule(() => deliver(payload)).catch(() => undefined);
-    // Start ordering stays global; each bridge still owns and drains its callback completion.
-    delivery = Promise.all([delivery, scheduled]).then(() => undefined);
-  });
-  return {
-    unsubscribe() {
-      if (unsubscribed) {
-        return;
-      }
-      unsubscribed = true;
-      rawUnsubscribe();
-    },
-    async drain(): Promise<void> {
-      await delivery;
-    },
-  };
-}
 
 type AgentEventBridge = {
   unsubscribe: () => void;
@@ -440,7 +365,7 @@ function createPlanUpdateBridge(params: {
     suppressed: params.suppressed,
     // GetReplyOptions callbacks may return void; the bridge awaits promises.
     deliver: deliver
-      ? async (payload) => {
+      ? async (payload: Parameters<NonNullable<GetReplyOptions["onPlanUpdate"]>>[0]) => {
           await deliver(payload);
         }
       : undefined,
